+++ src/controller_cli_callbacks.h
@@ -59,6 +59,8 @@ int cli_auto_merge_devs(clixon_handle h, cvec *cvv, cvec *argv);
 int cli_auto_del_devs(clixon_handle h, cvec *cvv, cvec *argv);
 int cli_auto_load_devs(clixon_handle h, cvec *cvv0, cvec *argv);
 int cli_controller_show_version(clixon_handle h, cvec *vars, cvec *argv);
+int show_yang_revisions(clixon_handle h, cvec *cvv, cvec *argv);
+int show_device_capability(clixon_handle h, cvec *cvv, cvec *argv);
 
 #ifdef __cplusplus
 }